# chore(telemetry): use importlibs instead of pkg_resources