--- a/ddtrace/internal/packages.py
+++ b/ddtrace/internal/packages.py
@@ -8,6 +8,7 @@ _DISTRIBUTIONS = None  # type: typing.Optional[typing.Set[Distribution]]
 
 def get_distributions():
     # type: () -> typing.Set[Distribution]
+    """returns the names, versions and path of all installed distributions"""
     global _DISTRIBUTIONS
     if _DISTRIBUTIONS is not None:
         return _DISTRIBUTIONS
@@ -19,11 +20,12 @@ def get_distributions():
 
     pkgs = set()
     for dist in importlib_metadata.distributions():
+        # accessing dist.metadata reads all lines in the PKG-INFO and/or METADATA files
+        # We should avoid accessing dist.metadata more than once
         metadata = dist.metadata
         name = metadata["name"]
         version = metadata["version"]
         if name and version:
-            # add distribution to a set to remove duplicates
             pkgs.add(Distribution(name, version))
 
     _DISTRIBUTIONS = pkgs