--- a/ddtrace/internal/telemetry/writer.py
+++ b/ddtrace/internal/telemetry/writer.py
@@ -18,6 +18,7 @@ from ..runtime import get_runtime_id
 from ..service import ServiceStatus
 from ..utils.time import StopWatch
 from .data import get_application
+from .data import get_dependencies
 from .data import get_host_info
 
 
@@ -183,12 +184,8 @@ class TelemetryWriter(PeriodicService):
         if self._forked:
             # app-started events should only be sent by the main process
             return
-        # pkg_resources import is inlined for performance reasons
-        # This import is an expensive operation
-        import pkg_resources
-
         payload = {
-            "dependencies": [{"name": pkg.project_name, "version": pkg.version} for pkg in pkg_resources.working_set],
+            "dependencies": get_dependencies(),
             "integrations": self._flush_integrations_queue(),
             "configurations": [],
         }