--- a/setup.py
+++ b/setup.py
@@ -246,6 +246,7 @@ setup(
         "attrs>=19.2.0",
         "six>=1.12.0",
         "typing_extensions; python_version<'3.8'",
+        "importlib_metadata; python_version<'3.8'",
     ]
     + bytecode,
     extras_require={