--- a/tests/internal/test_packages.py
+++ b/tests/internal/test_packages.py
@@ -1,23 +1,19 @@
-from ddtrace.internal.packages import Distribution
 from ddtrace.internal.packages import get_distributions
-from ddtrace.internal.utils.version import parse_version
 
 
 def test_get_distributions():
     """use pkg_resources to validate package names and versions returned by get_distributions()"""
     import pkg_resources
 
-    pkg_resources_ws = []
-    for pkg in pkg_resources.working_set:
-        # Rename typing-extension in working_set to match the package name returned
-        # by get_distributions()
-        if pkg.project_name == "typing-extensions" and parse_version(pkg.version) >= (4, 0):
-            pkg.project_name = "typing_extensions"
-        pkg_resources_ws.append(Distribution(pkg.project_name, pkg.version))
+    pkg_resources_ws = {pkg.project_name for pkg in pkg_resources.working_set}
 
-    importlib_pkg = get_distributions()
-    assert importlib_pkg
+    importlib_pkgs = {pkg.name for pkg in get_distributions()}
+    # The package name in typing_extensions-4.x.x.dist-info/METADATA is set to `typing_extensions`
+    # this is inconsistent with the package name found in pkg_resources. The block below corrects this.
+    # The correct package name is typing-extensions.
+    if "typing_extensions" in importlib_pkgs and "typing-extensions" in pkg_resources_ws:
+        importlib_pkgs.discard("typing_extensions")
+        importlib_pkgs.add("typing-extensions")
 
-    pkg_resources_ws.sort(key=lambda x: x.name)
-    importlib_pkg.sort(key=lambda x: x.name)
-    assert importlib_pkg == pkg_resources_ws
+    # assert that pkg_resources and importlib.metadata return the same packages
+    assert pkg_resources_ws == importlib_pkgs