--- a/tests/internal/test_packages.py
+++ b/tests/internal/test_packages.py
@@ -1,3 +1,5 @@
+import os
+
 from ddtrace.internal.packages import get_distributions
 
 
@@ -7,13 +9,18 @@ def test_get_distributions():
 
     pkg_resources_ws = {pkg.project_name for pkg in pkg_resources.working_set}
 
-    importlib_pkgs = {pkg.name for pkg in get_distributions()}
-    # The package name in typing_extensions-4.x.x.dist-info/METADATA is set to `typing_extensions`
-    # this is inconsistent with the package name found in pkg_resources. The block below corrects this.
-    # The correct package name is typing-extensions.
-    if "typing_extensions" in importlib_pkgs and "typing-extensions" in pkg_resources_ws:
-        importlib_pkgs.discard("typing_extensions")
-        importlib_pkgs.add("typing-extensions")
+    importlib_pkgs = set()
+    for pkg in get_distributions():
+        assert pkg.name
+        assert pkg.version
+        assert os.path.exists(pkg.path)
+        # The package name in typing_extensions-4.x.x.dist-info/METADATA is set to `typing_extensions`
+        # this is inconsistent with the package name found in pkg_resources. The block below corrects this.
+        # The correct package name is typing-extensions.
+        if pkg.name == "typing_extensions" and "typing-extensions" in pkg_resources_ws:
+            importlib_pkgs.add("typing-extensions")
+        else:
+            importlib_pkgs.add(pkg.name)
 
     # assert that pkg_resources and importlib.metadata return the same packages
     assert pkg_resources_ws == importlib_pkgs